feat: create a component for the google button auth
