+++ src/components/view/landing/TimelineLanding.tsx
@@ -89,7 +89,7 @@ const TimelineLanding: React.FC = () => {
               Neque Dolor, fugiat non cum doloribus aperiam voluptates nostrum.
                 </p>
               </div>
-              <a href='#' className='group-hover:text-secondary flex items-center justify-between'>
+              <a href='#' className='flex items-center justify-between group-hover:text-black'>
                 <span className='text-sm'>Read more</span>
                 <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='currentColor' className='h-5 w-5 -translate-x-4 text-2xl opacity-0 transition duration-300 group-hover:translate-x-0 group-hover:opacity-100'>
                   <path fillRule='evenodd' d='M12.97 3.97a.75.75 0 011.06 0l7.5 7.5a.75.75 0 010 1.06l-7.5 7.5a.75.75 0 11-1.06-1.06l6.22-6.22H3a.75.75 0 010-1.5h16.19l-6.22-6.22a.75.75 0 010-1.06z' clipRule='evenodd' />
