+++ src/views/welcome/WelcomeView.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
-import { useGoogleLogin } from '@react-oauth/google'
 import HeaderLanding from '../../components/view/landing/HeaderLanding'
 import DescriptionLanding from '../../components/view/landing/DescriptionLanding'
 import TimelineLanding from '../../components/view/landing/TimelineLanding'
@@ -8,35 +6,6 @@ import TeamLanding from '../../components/view/landing/TeamLanding'
 import ContactLanding from '../../components/view/landing/ContactLanding'
 
 const WelcomeView: React.FC = () => {
-  const login = useGoogleLogin({
-    onSuccess: tokenResponse => console.log(tokenResponse),
-  })
-
-  /*
-  const onGoogleSuccess = (e: any) => {
-    axios
-      .post(`${API_BASE}/auth/google/login`, {
-        token: e.tokenId,
-      })
-      .then((response) => {
-        (async () => login(response.data.accessToken))();
-        (async () => {
-          setRoute('/')
-        })()
-      })
-      .catch((err: AxiosError) => {
-        console.error(err)
-      })
-      .finally(() => {
-        setLoading(false)
-      })
-  }
-*/
-
-  const onGoogleFailure = () => {
-    // throw new Error("not implemented");
-  }
-  
   return (
     <div className='bg-white dark:bg-gray-900'>
       <HeaderLanding />
@@ -45,43 +14,6 @@ const WelcomeView: React.FC = () => {
         <TimelineLanding />
         <TeamLanding />
         <ContactLanding />
-        <div>
-          <div>Welcome to Noted!</div>
-          <Link to='/signin'>Go to the /signin page!</Link>
-          <br />  
-          <Link to='/signup'>Go to the /signup page!</Link>
-          <br />
-          <button className='dark:bg-gray-medium flex items-center justify-center rounded border border-gray-300 bg-white px-3 py-2 text-sm font-medium text-gray-800 dark:border-gray-500'
-            onClick={() => login()}>
-            <p className='mr-2 dark:text-gray-100'>Sign in with Google</p>
-            <svg
-              className='h-5 w-5'
-              xmlns='http://www.w3.org/2000/svg'
-              viewBox='0 0 24 24'
-              width='24'
-              height='24'
-            >
-              <g transform='matrix(1, 0, 0, 1, 27.009001, -39.238998)'>
-                <path
-                  fill='#4285F4'
-                  d='M -3.264 51.509 C -3.264 50.719 -3.334 49.969 -3.454 49.239 L -14.754 49.239 L -14.754 53.749 L -8.284 53.749 C -8.574 55.229 -9.424 56.479 -10.684 57.329 L -10.684 60.329 L -6.824 60.329 C -4.564 58.239 -3.264 55.159 -3.264 51.509 Z'
-                />
-                <path
-                  fill='#34A853'
-                  d='M -14.754 63.239 C -11.514 63.239 -8.804 62.159 -6.824 60.329 L -10.684 57.329 C -11.764 58.049 -13.134 58.489 -14.754 58.489 C -17.884 58.489 -20.534 56.379 -21.484 53.529 L -25.464 53.529 L -25.464 56.619 C -23.494 60.539 -19.444 63.239 -14.754 63.239 Z'
-                />
-                <path
-                  fill='#FBBC05'
-                  d='M -21.484 53.529 C -21.734 52.809 -21.864 52.039 -21.864 51.239 C -21.864 50.439 -21.724 49.669 -21.484 48.949 L -21.484 45.859 L -25.464 45.859 C -26.284 47.479 -26.754 49.299 -26.754 51.239 C -26.754 53.179 -26.284 54.999 -25.464 56.619 L -21.484 53.529 Z'
-                />
-                <path
-                  fill='#EA4335'
-                  d='M -14.754 43.989 C -12.984 43.989 -11.404 44.599 -10.154 45.789 L -6.734 42.369 C -8.804 40.429 -11.514 39.239 -14.754 39.239 C -19.444 39.239 -23.494 41.939 -25.464 45.859 L -21.484 48.949 C -20.534 46.099 -17.884 43.989 -14.754 43.989 Z'
-                />
-              </g>
-            </svg>
-          </button>
-        </div>
       </main>
     </div>
   )
